#33: Changes so heap_getChunk() returns 4-byte aligned chunks.

+++ src/vm/global.c
@@ -84,19 +84,19 @@ global_init(void)
     gVmGlobal.zero.od.od_type = OBJ_TYPE_INT;
     gVmGlobal.zero.od.od_size = sizeof(PmInt_t);
     gVmGlobal.zero.od.od_const = 1;
-    gVmGlobal.zero.val = 0;
+    gVmGlobal.zero.val = (int32_t)0;
 
     /* init one */
     gVmGlobal.one.od.od_type = OBJ_TYPE_INT;
     gVmGlobal.one.od.od_size = sizeof(PmInt_t);
     gVmGlobal.one.od.od_const = 1;
-    gVmGlobal.one.val = 1;
+    gVmGlobal.one.val = (int32_t)1;
 
     /* init negone */
     gVmGlobal.negone.od.od_type = OBJ_TYPE_INT;
     gVmGlobal.negone.od.od_size = sizeof(PmInt_t);
     gVmGlobal.negone.od.od_const = 1;
-    gVmGlobal.negone.val = (uint32_t)-1;
+    gVmGlobal.negone.val = (int32_t)-1;
 
     /* init None */
     gVmGlobal.none.od.od_type = OBJ_TYPE_NON;
